Add files via upload
In this paper, a novel reinforcement learning mission supervisor (RLMS) with memory is proposed for human-multi-robot coordination systems (HMRCS). The existing HMRCS are known to suffer from long decision waiting time and large mission error caused by repeated human intervention, restricting the autonomy of multi-robot systems. The proposed supervisor elaborately integrates deep-Q-network (DQN) and long-short-term memory (LSTM) knowledge base within the null-space based behavioral control (NSBC) framework, so as to achieve optimal adjustment strategy of the behavioral priority in the presence of mission conflicts, and to reduce the frequency of human intervention. In particular, the proposed RLMS with memory first memorize human intervention history when robot systems are not confident in decision making when encountering emergencies, and then reload the history information when encountering the same situation that have been tackled by human previously. Simulation demonstrates the effectiveness of proposed RLMS with memory.